refactor: Rename slackService -> monitoringService
