+++ apps/backend/src/routes/auth/strava.ts
@@ -8,7 +8,7 @@ import {
 import { isError, isSuccess } from '@dundring/utils';
 import * as express from 'express';
 import {
-  slackService,
+  monitoringService,
   stravaService,
   userService,
   validationService,
@@ -61,7 +61,7 @@ router.post<
     const fitnessData = await userService.getUserFitnessData(userId);
     const ftp = isSuccess(fitnessData)
       ? fitnessData.data.ftp
-      : slackService.logAndReturn(
+      : monitoringService.logAndReturn(
           `user [${userId}] does not have any fitnessData stored. Returning {ftp: 200}`,
           200
         );
