+++ apps/backend/src/services/index.ts
@@ -1,6 +1,6 @@
 import * as groupSessionService from './groupSessionService';
 import * as mailService from './mailService';
-import * as slackService from './slackService';
+import * as monitoringService from './monitoringService';
 import * as stravaService from './stravaService';
 import * as userService from './userService';
 import * as validationService from './validationService';
@@ -9,7 +9,7 @@ import * as workoutService from './workoutService';
 export {
   groupSessionService,
   mailService,
-  slackService,
+  monitoringService,
   stravaService,
   userService,
   validationService,
